refactor: break up BrowserUnit and HelperUnit into focused helpers
BrowserUnit (896→379 lines, -58%):
- DownloadHelper (299 lines): download, saveDataUrl, guessFilename, etc.
- UrlHelper (225 lines): isURL, queryWrapper, stripUrlQuery, neatUrl
- BookmarkRenderer (98 lines): loadRecentlyUsedBookmarks, getRecentBookmarksContent

HelperUnit (875→365 lines, -58%):
- MarkdownParser (217 lines): parseMarkdown, textMarkDown
- AppUpdater (238 lines): upgradeToLatestRelease, installApkFromFile
- FileHelper (116 lines): loadAssetFile, loadAssetFileToString, getFileInfoFromContentUri

Both retain forwarding stubs so existing callers need no import changes.
