+++ app/src/main/java/info/plateaukao/einkbro/unit/BrowserUnit.kt
@@ -2,12 +2,9 @@ package info.plateaukao.einkbro.unit
 
 import android.annotation.SuppressLint
 import android.app.Activity
-import android.app.DownloadManager
 import android.content.BroadcastReceiver
 import android.content.Context
-import android.content.Context.DOWNLOAD_SERVICE
 import android.content.Intent
-import android.content.Intent.ACTION_VIEW
 import android.content.pm.ShortcutManager
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
@@ -17,10 +14,6 @@ import android.os.Environment
 import android.os.Handler
 import android.os.Looper
 import android.os.Message
-import android.util.Log
-import android.webkit.CookieManager
-import android.webkit.URLUtil
-import android.webkit.WebSettings
 import android.webkit.WebView
 import android.webkit.WebView.HitTestResult.ANCHOR_TYPE
 import android.webkit.WebView.HitTestResult.IMAGE_ANCHOR_TYPE
@@ -35,42 +28,26 @@ import androidx.annotation.RequiresApi
 import androidx.appcompat.view.ContextThemeWrapper
 import androidx.core.app.ActivityCompat.finishAffinity
 import androidx.fragment.app.FragmentActivity
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.lifecycleScope
 import info.plateaukao.einkbro.R
 import info.plateaukao.einkbro.browser.Cookie
 import info.plateaukao.einkbro.database.RecordDb
 import info.plateaukao.einkbro.preference.ConfigManager
 import info.plateaukao.einkbro.preference.CustomFontInfo
-import info.plateaukao.einkbro.unit.HelperUnit.needGrantStoragePermission
 import info.plateaukao.einkbro.util.Constants
 import info.plateaukao.einkbro.view.EBToast
-import info.plateaukao.einkbro.view.EBToast.showShort
 import info.plateaukao.einkbro.view.EBWebView
-import info.plateaukao.einkbro.view.dialog.DialogManager
 import info.plateaukao.einkbro.view.dialog.ShortcutEditDialog
-import info.plateaukao.einkbro.view.dialog.TextInputDialog
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import org.json.JSONObject
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import java.io.ByteArrayInputStream
 import java.io.File
 import java.io.IOException
 import java.io.InputStream
-import java.io.UnsupportedEncodingException
-import java.net.HttpURLConnection
-import java.net.URL
-import java.net.URLDecoder
-import java.net.URLEncoder
-import java.util.Base64
-import java.util.Locale
 import java.util.Objects
-import java.util.regex.Pattern
 import kotlin.system.exitProcess
 
 @OptIn(DelicateCoroutinesApi::class)
@@ -80,87 +57,61 @@ object BrowserUnit : KoinComponent {
     private const val SUFFIX_TXT = ".txt"
     const val MIME_TYPE_TEXT_PLAIN = "text/plain"
     const val MIME_TYPE_IMAGE = "image/png"
-    private const val SEARCH_ENGINE_GOOGLE = "https://www.google.com/search?q="
-    private const val SEARCH_ENGINE_DUCKDUCKGO = "https://duckduckgo.com/?q="
-    private const val SEARCH_ENGINE_STARTPAGE = "https://startpage.com/do/search?query="
-    private const val SEARCH_ENGINE_BING = "http://www.bing.com/search?q="
-    private const val SEARCH_ENGINE_BAIDU = "https://www.baidu.com/s?wd="
-    private const val SEARCH_ENGINE_QWANT = "https://www.qwant.com/?q="
-    private const val SEARCH_ENGINE_ECOSIA = "https://www.ecosia.org/search?q="
-    private const val SEARCH_ENGINE_YANDEX = "https://yandex.com/search/?text="
-    private const val SEARCH_ENGINE_STARTPAGE_DE =
-        "https://startpage.com/do/search?lui=deu&language=deutsch&query="
-    private const val SEARCH_ENGINE_SEARX = "https://searx.me/?q="
     const val URL_ENCODING = "utf-8"
     const val URL_ABOUT_BLANK = "about:blank"
     const val URL_SCHEME_ABOUT = "about:"
     const val URL_SCHEME_MAIL_TO = "mailto:"
-    private const val URL_SCHEME_FILE = "file://"
-    private const val URL_SCHEME_HTTPS = "https://"
-    private const val URL_SCHEME_HTTP = "http://"
     const val URL_SCHEME_INTENT = "intent://"
-    private const val URL_PREFIX_GOOGLE_PLAY = "www.google.com/url?q="
-    private const val URL_SUFFIX_GOOGLE_PLAY = "&sa"
     val UA_DESKTOP_PREFIX = "Mozilla/5.0 (X11; Linux " + System.getProperty("os.arch") + ")"
     val UA_MOBILE_PREFIX = "Mozilla/5.0 (Linux; Android " + Build.VERSION.RELEASE + ")"
 
     private val config: ConfigManager by inject()
     val cookie: Cookie by inject()
 
-    private val neatUrlConfigs: List<String> = parseNeatUrlConfigs()
+    // --- Forwarding stubs for DownloadHelper (preserves existing call sites) ---
 
-    private fun dataUrlToMimeType(dataUrl: String): String =
-        dataUrl.substring(dataUrl.indexOf("/") + 1, dataUrl.indexOf(";"))
+    @JvmStatic
+    fun download(context: Context, url: String, contentDisposition: String, mimeType: String) =
+        DownloadHelper.download(context, url, contentDisposition, mimeType)
+
+    var downloadFileId: Long
+        get() = DownloadHelper.downloadFileId
+        set(value) { DownloadHelper.downloadFileId = value }
+
+    fun createDownloadReceiver(activity: Activity): BroadcastReceiver =
+        DownloadHelper.createDownloadReceiver(activity)
+
+    fun openDownloadFolder(activity: Activity) =
+        DownloadHelper.openDownloadFolder(activity)
+
+    // --- Forwarding stubs for UrlHelper (preserves existing call sites) ---
+
+    @JvmStatic
+    fun isURL(url: String?): Boolean = UrlHelper.isURL(url)
+
+    fun queryWrapper(context: Context, query: String): String =
+        UrlHelper.queryWrapper(context, query)
+
+    fun stripUrlQuery(url: String): String = UrlHelper.stripUrlQuery(url)
 
     @RequiresApi(Build.VERSION_CODES.O)
-    fun dataUrlToStream(dataUrl: String): InputStream {
-        val data = dataUrl.split(",")
-        val base64 = data[1]
-        val imageBytes = Base64.getDecoder().decode(base64)
-        return ByteArrayInputStream(imageBytes)
-    }
+    fun dataUrlToStream(dataUrl: String): InputStream = UrlHelper.dataUrlToStream(dataUrl)
 
-    fun createDownloadReceiver(activity: Activity): BroadcastReceiver {
-        return object : BroadcastReceiver() {
-            override fun onReceive(context: Context, intent: Intent) {
-                if (activity.isFinishing || downloadFileId == -1L) return
+    // --- Forwarding stubs for BookmarkRenderer (preserves existing call sites) ---
 
-                val downloadManager = activity.getSystemService(DOWNLOAD_SERVICE) as DownloadManager
-                val mostRecentDownload: Uri =
-                    downloadManager.getUriForDownloadedFile(downloadFileId) ?: return
-                val mimeType: String = downloadManager.getMimeTypeForDownloadedFile(downloadFileId)
-                val fileIntent = Intent(ACTION_VIEW).apply {
-                    setDataAndType(mostRecentDownload, mimeType)
-                    addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                    addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-                }
-                downloadFileId = -1L
-                DialogManager(activity).showOkCancelDialog(
-                    messageResId = R.string.toast_downloadComplete,
-                    okAction = {
-                        try {
-                            activity.startActivity(fileIntent)
-                        } catch (e: Exception) {
-                            e.printStackTrace()
-                            EBToast.show(activity, R.string.toast_error)
-                        }
-                    }
-                )
-            }
-        }
-    }
+    fun loadRecentlyUsedBookmarks(webView: EBWebView) =
+        BookmarkRenderer.loadRecentlyUsedBookmarks(webView)
 
-    fun openDownloadFolder(activity: Activity) {
-        val uri = Uri.parse(
-            Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-                .toString()
-        )
-        val intent =
-            Intent(Intent.ACTION_GET_CONTENT).apply { setDataAndType(uri, "resource/folder") }
-        activity.startActivity(
-            Intent.createChooser(intent, activity.getString(R.string.dialog_title_download))
-        )
-    }
+    fun getRecentBookmarksContent(context: Context): String =
+        BookmarkRenderer.getRecentBookmarksContent(context)
+
+    suspend fun getResourceAndMimetypeFromUrl(url: String, timeout: Int = 0): Pair<ByteArray, String> =
+        BookmarkRenderer.getResourceAndMimetypeFromUrl(url, timeout)
+
+    suspend fun getResourceFromUrl(url: String, timeout: Int = 0): ByteArray =
+        BookmarkRenderer.getResourceFromUrl(url, timeout)
+
+    // --- Remaining functions ---
 
     fun getWebViewLinkImageUrl(webView: WebView, message: Message): String {
         val hitTestResult = webView.hitTestResult
@@ -225,7 +176,7 @@ object BrowserUnit : KoinComponent {
                     super.onPageFinished(view, url)
                     view?.let {
                         view.evaluateJavascript(
-                            """ 
+                            """
                             document.getElementById("bookmark").remove();
                             document.getElementsByClassName("gnb_wrap")[0].remove();
                             document.getElementsByClassName("search_wrap")[0].remove();
@@ -236,7 +187,7 @@ object BrowserUnit : KoinComponent {
                     view?.postDelayed(
                         {
                             view.evaluateJavascript(
-                                """ 
+                                """
                             document.getElementById("_id_mobile_ad").remove();
                         """.trimIndent(), null
                             )
@@ -250,268 +201,6 @@ object BrowserUnit : KoinComponent {
     }
 
     @JvmStatic
-    fun isURL(url: String?): Boolean {
-        var url = url ?: return false
-        url = url.lowercase(Locale.getDefault())
-        if (url.startsWith(URL_ABOUT_BLANK)
-            || url.startsWith(URL_SCHEME_MAIL_TO)
-            || url.startsWith(URL_SCHEME_FILE)
-        ) {
-            return true
-        }
-        val regex = ("^((ftp|http|https|intent)?://)" // support scheme
-                + "?(([0-9a-z_!~*'().&=+$%-]+: )?[0-9a-z_!~*'().&=+$%-]+@)?" // ftp的user@
-                + "(([0-9]{1,3}\\.){3}[0-9]{1,3}" // IP形式的URL -> 199.194.52.184
-                + "|" // 允许IP和DOMAIN（域名）
-                + "(.)*" // 域名 -> www.
-                // + "([0-9a-z_!~*'()-]+\\.)*"                               // 域名 -> www.
-                + "([0-9a-z][0-9a-z-]{0,61})?[0-9a-z]\\." // 二级域名
-                + "[a-z]{2,6})" // first level domain -> .com or .museum
-                + "(:[0-9]{1,4})?" // 端口 -> :80
-                + "((/?)|" // a slash isn't required if there is no file name
-                + "(/[0-9a-z_!~*'().;?:@&=+$,%#-]+)+/?)$")
-        val pattern = Pattern.compile(regex)
-        val isMatch = pattern.matcher(url).matches()
-        return if (isMatch) true else try {
-            val uri = Uri.parse(url)
-            val scheme = uri.scheme
-            scheme == "ftp" || scheme == "http" || scheme == "https" || scheme == "intent"
-        } catch (exception: Exception) {
-            false
-        }
-    }
-
-    fun queryWrapper(context: Context, query: String): String {
-        // Use prefix and suffix to process some special links
-        var query = query
-        val temp = query.lowercase(Locale.getDefault())
-        if (temp.contains(URL_PREFIX_GOOGLE_PLAY) && temp.contains(URL_SUFFIX_GOOGLE_PLAY)) {
-            val start = temp.indexOf(URL_PREFIX_GOOGLE_PLAY) + URL_PREFIX_GOOGLE_PLAY.length
-            val end = temp.indexOf(URL_SUFFIX_GOOGLE_PLAY)
-            query = query.substring(start, end)
-        }
-
-        // remove prefix non-url part
-        if (config.shouldTrimInputUrl) {
-            var foundIndex = query.indexOf(URL_SCHEME_HTTPS)
-            if (foundIndex > 0) {
-                query = query.substring(foundIndex)
-            }
-            foundIndex = query.indexOf(URL_SCHEME_HTTP)
-            if (foundIndex > 0) {
-                query = query.substring(foundIndex)
-            }
-        }
-
-        if (isURL(query)) {
-            if (query.startsWith(URL_SCHEME_ABOUT) || query.startsWith(URL_SCHEME_MAIL_TO)) {
-                return query
-            }
-            if (!query.contains("://")) {
-                query = URL_SCHEME_HTTPS + query
-            }
-            return query.replace(" ", "+")
-        }
-
-        try {
-            query = URLEncoder.encode(query, URL_ENCODING)
-        } catch (u: UnsupportedEncodingException) {
-            Log.w("browser", "Unsupported Encoding Exception")
-        }
-        return when (config.searchEngine.toInt()) {
-            0 -> SEARCH_ENGINE_STARTPAGE + query
-            1 -> SEARCH_ENGINE_STARTPAGE_DE + query
-            2 -> SEARCH_ENGINE_BAIDU + query
-            3 -> SEARCH_ENGINE_BING + query
-            6 -> SEARCH_ENGINE_SEARX + query
-            7 -> SEARCH_ENGINE_QWANT + query
-            8 -> SEARCH_ENGINE_ECOSIA + query
-            9 -> config.searchEngineUrl + query
-            10 -> SEARCH_ENGINE_YANDEX + query
-            5 -> SEARCH_ENGINE_GOOGLE + query
-            4 -> SEARCH_ENGINE_DUCKDUCKGO + query
-            else -> SEARCH_ENGINE_GOOGLE + query
-        }
-    }
-
-    fun bitmap2File(context: Context, bitmap: Bitmap, filename: String?): Boolean {
-        try {
-            val fileOutputStream = context.openFileOutput(filename, Context.MODE_PRIVATE)
-            bitmap.compress(Bitmap.CompressFormat.PNG, 100, fileOutputStream)
-            fileOutputStream.flush()
-            fileOutputStream.close()
-        } catch (e: Exception) {
-            return false
-        }
-        return true
-    }
-
-    fun file2Bitmap(context: Context, filename: String?): Bitmap? {
-        return try {
-            val fileInputStream = context.openFileInput(filename)
-            BitmapFactory.decodeStream(fileInputStream)
-        } catch (e: Exception) {
-            null
-        }
-    }
-
-    @JvmStatic
-    fun download(context: Context, url: String, contentDisposition: String, mimeType: String) {
-        val activity = context as Activity
-        if (needGrantStoragePermission(activity)) {
-            return
-        }
-
-        if (url.startsWith("data:")) {
-            saveDataUrl(activity, url, mimeType)
-            return
-        }
-
-        val filename = Uri.decode(guessFilename(url, contentDisposition, mimeType))
-
-        if (hasExactFilename(contentDisposition)) {
-            internalDownload(activity, url, mimeType, filename)
-        } else {
-            val title = context.getString(R.string.dialog_title_download)
-            (activity as LifecycleOwner).lifecycleScope.launch {
-                val modifiedFilename = TextInputDialog(
-                    context,
-                    "",
-                    title,
-                    filename
-                ).show()
-
-                modifiedFilename?.let { internalDownload(activity, url, mimeType, it) }
-            }
-        }
-    }
-
-    private fun saveDataUrl(activity: Activity, dataUrl: String, fallbackMimeType: String) {
-        try {
-            val header = dataUrl.substring(dataUrl.indexOf(":") + 1, dataUrl.indexOf(","))
-            val dataMimeType = if (header.contains(";")) {
-                header.substring(0, header.indexOf(";"))
-            } else {
-                header
-            }
-            val isBase64 = header.contains("base64", ignoreCase = true)
-            val rawData = dataUrl.substring(dataUrl.indexOf(",") + 1)
-
-            val effectiveMimeType = dataMimeType.ifEmpty { fallbackMimeType }
-            val ext = android.webkit.MimeTypeMap.getSingleton()
-                .getExtensionFromMimeType(effectiveMimeType) ?: "bin"
-            val filename = "download_${System.currentTimeMillis()}.$ext"
-
-            val downloadsDir =
-                Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-            val destFile = File(downloadsDir, filename)
-
-            val bytes = if (isBase64) {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                    Base64.getDecoder().decode(rawData)
-                } else {
-                    android.util.Base64.decode(rawData, android.util.Base64.DEFAULT)
-                }
-            } else {
-                java.net.URLDecoder.decode(rawData, "UTF-8").toByteArray()
-            }
-
-            destFile.writeBytes(bytes)
-            showShort(activity, R.string.toast_downloadComplete)
-        } catch (e: Exception) {
-            Log.w("browser", "Failed to save data URL: $e")
-            showShort(activity, R.string.error_download_link_invalid)
-        }
-    }
-
-    private fun hasExactFilename(contentDisposition: String): Boolean =
-        parseContentDispositionFilename(contentDisposition) != null
-
-    var downloadFileId = -1L
-    private fun internalDownload(
-        activity: Activity,
-        url: String,
-        mimeType: String,
-        filename: String
-    ) {
-        val cookie = CookieManager.getInstance().getCookie(url)
-        val userAgent = WebSettings.getDefaultUserAgent(activity)
-        if (Uri.parse(url).host == null) {
-            showShort(activity, R.string.error_download_link_invalid)
-            return
-        }
-        try {
-            val request = DownloadManager.Request(Uri.parse(url)).apply {
-                if (cookie != null) addRequestHeader("Cookie", cookie)
-                addRequestHeader("User-Agent", userAgent)
-                setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
-                setTitle(filename)
-                setMimeType(mimeType)
-                try {
-                    setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, filename)
-                } catch (e: Exception) {
-                    val downloadsDir =
-                        Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-                    setDestinationUri(Uri.fromFile(java.io.File(downloadsDir, filename)))
-                }
-            }
-            val manager = activity.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
-            downloadFileId = manager.enqueue(request)
-            showShort(activity, R.string.toast_start_download)
-        } catch (e: Exception) {
-            Log.w("browser", "DownloadManager failed, falling back to direct download: $e")
-            directDownload(activity, url, cookie, userAgent, filename, mimeType)
-        }
-    }
-
-    private fun directDownload(
-        activity: Activity,
-        url: String,
-        cookie: String?,
-        userAgent: String,
-        filename: String,
-        mimeType: String
-    ) {
-        showShort(activity, R.string.toast_start_download)
-        (activity as LifecycleOwner).lifecycleScope.launch(Dispatchers.IO) {
-            try {
-                val connection = URL(url).openConnection() as HttpURLConnection
-                connection.instanceFollowRedirects = true
-                connection.addRequestProperty("User-Agent", userAgent)
-                if (cookie != null) connection.addRequestProperty("Cookie", cookie)
-                connection.connectTimeout = 15000
-                connection.readTimeout = 30000
-                connection.connect()
-
-                if (connection.responseCode != HttpURLConnection.HTTP_OK) {
-                    withContext(Dispatchers.Main) {
-                        showShort(activity, R.string.error_download_link_invalid)
-                    }
-                    return@launch
-                }
-
-                val downloadsDir =
-                    Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-                val destFile = File(downloadsDir, filename)
-                connection.inputStream.use { input ->
-                    destFile.outputStream().use { output ->
-                        input.copyTo(output)
-                    }
-                }
-
-                withContext(Dispatchers.Main) {
-                    showShort(activity, R.string.toast_downloadComplete)
-                }
-            } catch (e: Exception) {
-                Log.w("browser", "Direct download also failed: $e")
-                withContext(Dispatchers.Main) {
-                    showShort(activity, R.string.error_download_link_invalid)
-                }
-            }
-        }
-    }
-
-    @JvmStatic
     fun clearCache(context: Context) {
         try {
             val dir = context.cacheDir
@@ -519,13 +208,13 @@ object BrowserUnit : KoinComponent {
                 deleteDir(dir)
             }
         } catch (exception: Exception) {
-            Log.w("browser", "Error clearing cache")
+            android.util.Log.w("browser", "Error clearing cache")
         }
     }
 
     @JvmStatic
     fun clearCookie() {
-        val cookieManager = CookieManager.getInstance()
+        val cookieManager = android.webkit.CookieManager.getInstance()
         cookieManager.flush()
         cookieManager.removeAllCookies { }
     }
@@ -565,71 +254,26 @@ object BrowserUnit : KoinComponent {
         return dir != null && dir.delete()
     }
 
-    fun guessFilename(url: String, contentDisposition: String, mimeType: String): String {
-        val cdFilename = parseContentDispositionFilename(contentDisposition)
-        if (cdFilename != null) return sanitizeFilename(cdFilename)
-
-        val urlFilename = extractFilenameFromUrl(url, mimeType)
-        if (urlFilename != null) return sanitizeFilename(urlFilename)
-
-        return sanitizeFilename(URLUtil.guessFileName(url, contentDisposition, mimeType))
-    }
-
-    private fun parseContentDispositionFilename(contentDisposition: String): String? {
-        if (contentDisposition.isBlank()) return null
-
-        // RFC 5987: filename*=UTF-8''percent-encoded-name (takes priority per RFC 6266)
-        val extValuePattern = Regex("""filename\*\s*=\s*[^']*'[^']*'(.+?)(?:\s*;|$)""", RegexOption.IGNORE_CASE)
-        extValuePattern.find(contentDisposition)?.let { match ->
-            val encoded = match.groupValues[1].trim()
-            return try {
-                URLDecoder.decode(encoded, "UTF-8")
-            } catch (e: Exception) {
-                encoded
-            }
+    fun bitmap2File(context: Context, bitmap: Bitmap, filename: String?): Boolean {
+        try {
+            val fileOutputStream = context.openFileOutput(filename, Context.MODE_PRIVATE)
+            bitmap.compress(Bitmap.CompressFormat.PNG, 100, fileOutputStream)
+            fileOutputStream.flush()
+            fileOutputStream.close()
+        } catch (e: Exception) {
+            return false
         }
+        return true
+    }
 
-        // Quoted: filename="name"
-        val quotedPattern = Regex("""filename\s*=\s*"([^"]+)"""", RegexOption.IGNORE_CASE)
-        quotedPattern.find(contentDisposition)?.let { match ->
-            return match.groupValues[1]
+    fun file2Bitmap(context: Context, filename: String?): Bitmap? {
+        return try {
+            val fileInputStream = context.openFileInput(filename)
+            BitmapFactory.decodeStream(fileInputStream)
+        } catch (e: Exception) {
+            null
         }
-
-        // Unquoted: filename=name
-        val unquotedPattern = Regex("""filename\s*=\s*([^\s;]+)""", RegexOption.IGNORE_CASE)
-        unquotedPattern.find(contentDisposition)?.let { match ->
-            return match.groupValues[1]
-        }
-
-        return null
     }
-
-    private fun extractFilenameFromUrl(url: String, mimeType: String): String? {
-        val decodedUrl = try { Uri.decode(url) } catch (e: Exception) { return null } ?: return null
-
-        // Strip query string and fragment
-        val cleanUrl = decodedUrl.substringBefore('?').substringBefore('#')
-        if (cleanUrl.endsWith("/")) return null
-
-        val lastSegment = cleanUrl.substringAfterLast('/')
-        if (lastSegment.isBlank()) return null
-
-        // If it has a recognizable file extension, use it
-        if (lastSegment.contains('.')) return lastSegment
-
-        // No extension — try to derive one from mimeType
-        val ext = android.webkit.MimeTypeMap.getSingleton().getExtensionFromMimeType(mimeType)
-        return if (ext != null) "$lastSegment.$ext" else null
-    }
-
-    private fun sanitizeFilename(filename: String): String {
-        return filename
-            .replace(Regex("[/\\\\]"), "_")
-            .replace(Regex("[\\x00-\\x1f]"), "")
-            .trim()
-            .ifBlank { "download" }
-    }
-
 
     fun openFontFilePicker(resultLauncher: ActivityResultLauncher<Intent>) {
         val intent = Intent(Intent.ACTION_OPEN_DOCUMENT)
@@ -655,8 +299,8 @@ object BrowserUnit : KoinComponent {
 
     @RequiresApi(Build.VERSION_CODES.O)
     fun saveImageFromUrl(url: String, resultLauncher: ActivityResultLauncher<Intent>) {
-        val fileFormat = dataUrlToMimeType(url)
-        tempImageInputStream = dataUrlToStream(url)
+        val fileFormat = UrlHelper.dataUrlToMimeType(url)
+        tempImageInputStream = UrlHelper.dataUrlToStream(url)
         val mimeType = when (fileFormat.lowercase()) {
             "png" -> "image/png"
             "jpg" -> "image/jpeg"
@@ -697,102 +341,6 @@ object BrowserUnit : KoinComponent {
         }
     }
 
-    fun stripUrlQuery(url: String): String {
-        if (!config.shouldPruneQueryParameters) return url
-
-        try {
-            var strippedCount = 0
-            val uri = Uri.parse(url)
-            if (uri.authority == null) return url
-
-            val params = uri.queryParameterNames
-            if (params.isEmpty()) return url
-
-            val uriBuilder = uri.buildUpon().clearQuery()
-            for (param in params) {
-                if (!matchNeatUrlConfig(uri.host.orEmpty(), param)) {
-                    uriBuilder.appendQueryParameter(param, uri.getQueryParameter(param))
-                } else {
-                    strippedCount++
-                }
-            }
-
-            // only return the stripped url if we stripped something
-            // to fix the only key but no value scenario
-            return if (strippedCount > 0) {
-                Log.d("strippedCount", "$strippedCount")
-                uriBuilder.build().toString()
-            } else {
-                url
-            }
-        } catch (e: Exception) {
-            return url
-        }
-    }
-
-    private fun matchNeatUrlConfig(host: String, param: String): Boolean {
-        neatUrlConfigs.forEach { paramConfig ->
-            // handle host part
-            if (paramConfig.contains("@")) {
-                val paramConfigs = paramConfig.split("@")
-                if (paramConfigs.size == 2) {
-                    val paramValue = paramConfigs[0]
-                    val hostValue = paramConfigs[1]
-                    val modifiedHost = host.replace("www.", "")
-                    if (matchStarString(hostValue, modifiedHost) &&
-                        matchStarString(paramValue, param)
-                    ) {
-                        return true
-                    }
-                }
-            }
-
-            // handle normal param part
-            if (matchStarString(paramConfig, param)) return true
-        }
-        return false
-    }
-    
-    fun loadRecentlyUsedBookmarks(webView: EBWebView) {
-        val html = getRecentBookmarksContent(webView.context)
-        if (html.isNotBlank()) {
-            webView.loadDataWithBaseURL(
-                null,
-                html,
-                "text/html",
-                "utf-8",
-                null
-            )
-            webView.albumTitle = webView.context.getString(R.string.recently_used_bookmarks)
-        }
-    }
-
-    private fun matchStarString(config: String, param: String): Boolean {
-        if (config.endsWith("*")) {
-            if (param.startsWith(config.substring(0, config.length - 1))) return true
-        } else if (config.startsWith("*")) {
-            if (param.endsWith(config.substring(1, config.length))) return true
-        } else if (config == param) {
-            return true
-        }
-        return false
-    }
-
-    internal data class NeatUrlConfig(val name: String, val params: List<String>)
-
-    @Suppress("UNCHECKED_CAST")
-    private fun parseNeatUrlConfigs(): List<String> {
-        val configArray = JSONObject(Constants.NEAT_URL_DATA)
-            .getJSONArray("categories")
-
-        return (0 until configArray.length()).map { index ->
-            val config = configArray.getJSONObject(index)
-            val paramsArray = config.getJSONArray("params")
-            (0 until paramsArray.length()).map { s -> paramsArray.getString(s) }
-        }.flatten()
-    }
-
-
     fun handleSaveImageFilePickerResult(
         activity: ComponentActivity,
         activityResult: ActivityResult,
@@ -804,69 +352,6 @@ object BrowserUnit : KoinComponent {
         tempImageInputStream?.let { saveImage(activity, it, uri, postAction) }
         tempImageInputStream?.close()
         tempImageInputStream = null
-    }
-
-    fun getRecentBookmarksContent(context: Context): String {
-        if (config.recentBookmarks.isEmpty()) return ""
-        val alignBottom = !config.isToolbarOnTop
-        val bookmarks = if (alignBottom) config.recentBookmarks.reversed() else config.recentBookmarks
-        val content = bookmarks.joinToString(separator = "\n") {
-            val initial = it.name.firstOrNull()?.uppercase() ?: "#"
-            val domain = try {
-                java.net.URI(it.url).host?.removePrefix("www.") ?: ""
-            } catch (e: Exception) { "" }
-            val faviconUrl = try {
-                val uri = java.net.URI(it.url)
-                "${uri.scheme}://${uri.host}/favicon.ico"
-            } catch (e: Exception) { "" }
-            """
-            <a href="${it.url}" class="card">
-                <div class="icon">
-                    <img src="$faviconUrl" onerror="this.style.display='none';this.nextElementSibling.style.display='flex'" />
-                    <span class="fallback">$initial</span>
-                </div>
-                <div class="info">
-                    <div class="name">${it.name}</div>
-                    <div class="domain">$domain</div>
-                </div>
-            </a>
-            """
-        }
-        val bodyClass = if (alignBottom) "align-bottom" else ""
-        return HelperUnit.loadAssetFileToString(context, "recent_bookmarks.html")
-            .replace("{{BODY_CLASS}}", bodyClass)
-            .replace("{{CONTENT}}", content)
-    }
-
-    suspend fun getResourceAndMimetypeFromUrl(url: String, timeout: Int = 0): Pair<ByteArray, String> {
-        var byteArray: ByteArray = "".toByteArray()
-        var mimeType = ""
-        withContext(Dispatchers.IO) {
-            try {
-                val connection: HttpURLConnection = URL(url).openConnection() as HttpURLConnection
-                connection.addRequestProperty("User-Agent", "Mozilla/4.76")
-                if (timeout > 0) connection.connectTimeout = timeout
-                connection.connect()
-                if (connection.responseCode != HttpURLConnection.HTTP_OK) {
-                    if (isRedirect(connection.responseCode)) {
-                        val redirectUrl = connection.getHeaderField("Location")
-                        byteArray = getResourceFromUrl(redirectUrl, timeout)
-                    }
-                } else {
-                    mimeType = connection.contentType
-                    byteArray = connection.inputStream.readBytes()
-                    connection.inputStream.close()
-                }
-            } catch (e: IOException) {
-                Log.w("browser", "Failed getting resource: $e")
-                e.printStackTrace()
-            }
-        }
-        return Pair(byteArray, mimeType)
-    }
-
-    suspend fun getResourceFromUrl(url: String, timeout: Int = 0): ByteArray {
-        return getResourceAndMimetypeFromUrl(url, timeout).first
     }
 
     fun createShortcut(activity: FragmentActivity, ebWebView: EBWebView) {
@@ -891,7 +376,4 @@ object BrowserUnit : KoinComponent {
         activity.overridePendingTransition(0, 0)
         exitProcess(0)
     }
-
-
-    private fun isRedirect(responseCode: Int): Boolean = responseCode in 301..399
 }
