+++ app/src/main/java/info/plateaukao/einkbro/unit/HelperUnit.kt
@@ -30,61 +30,63 @@ import android.graphics.Bitmap
 import android.graphics.drawable.Icon
 import android.net.Uri
 import android.os.Build
-import android.provider.OpenableColumns
-import android.util.Log
 import android.view.View
-import android.webkit.MimeTypeMap
 import androidx.activity.result.ActivityResultLauncher
 import androidx.annotation.RequiresApi
 import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.unit.TextUnit
-import androidx.compose.ui.unit.sp
-import androidx.core.content.FileProvider
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetBehavior.BottomSheetCallback
 import com.google.android.material.bottomsheet.BottomSheetDialog
-import info.plateaukao.einkbro.BuildConfig
-import info.plateaukao.einkbro.EinkBroApplication
 import info.plateaukao.einkbro.R
 import info.plateaukao.einkbro.activity.EpubReaderActivity
 import info.plateaukao.einkbro.util.Constants
 import info.plateaukao.einkbro.view.EBToast
 import info.plateaukao.einkbro.view.dialog.DialogManager
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import org.json.JSONArray
-import java.io.BufferedReader
-import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import java.io.InputStream
-import java.io.InputStreamReader
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
-import java.util.zip.ZipInputStream
 
 
 object HelperUnit {
     private const val REQUEST_CODE_ASK_PERMISSIONS = 123
     private const val REQUEST_CODE_ASK_PERMISSIONS_1 = 1234
 
-    private val client = OkHttpClient.Builder()
-        .followRedirects(true)
-        .followSslRedirects(true)
-        .build()
+    // --- Forwarding functions for MarkdownParser ---
+    fun parseMarkdown(markdownText: String): AnnotatedString =
+        MarkdownParser.parseMarkdown(markdownText)
 
+    // --- Forwarding functions for AppUpdater ---
+    suspend fun upgradeToLatestRelease(
+        context: Context,
+        progressCallback: (suspend (Float) -> Unit)? = null,
+    ) = AppUpdater.upgradeToLatestRelease(context, progressCallback)
 
+    suspend fun upgradeFromSnapshot(
+        context: Context,
+        progressCallback: (suspend (Float) -> Unit)? = null,
+    ) = AppUpdater.upgradeFromSnapshot(context, progressCallback)
+
+    // --- Forwarding functions for FileHelper ---
     fun getStringFromAsset(fileName: String): String =
-        EinkBroApplication.instance.assets.open(fileName).bufferedReader().use { it.readText() }
+        FileHelper.getStringFromAsset(fileName)
+
+    fun getCachedPathFromURI(context: Context, contentURI: Uri): String =
+        FileHelper.getCachedPathFromURI(context, contentURI)
+
+    fun readContentAsStringList(contentResolver: ContentResolver, contentUri: Uri): List<String> =
+        FileHelper.readContentAsStringList(contentResolver, contentUri)
+
+    fun getFileInfoFromContentUri(context: Context, contentUri: Uri): Pair<String?, String?> =
+        FileHelper.getFileInfoFromContentUri(context, contentUri)
+
+    fun loadAssetFileToString(context: Context, filename: String): String =
+        FileHelper.loadAssetFileToString(context, filename)
+
+    fun loadAssetFile(fileName: String): String =
+        FileHelper.loadAssetFile(fileName)
+
+    @JvmStatic
+    fun fileName(url: String?): String =
+        FileHelper.fileName(url)
+
+    // --- Remaining functions ---
 
     @JvmStatic
     // return true if need permissions
@@ -205,33 +207,6 @@ object HelperUnit {
         return buildUpon().scheme(scheme).build()
     }
 
-    fun getCachedPathFromURI(context: Context, contentURI: Uri): String {
-        //val tempFile = File(context.filesDir.absolutePath + "/temp.mht")
-        val tempFile = File.createTempFile("tempfile", ".mht", context.cacheDir)
-        tempFile.deleteOnExit()
-
-        context.contentResolver.openInputStream(contentURI)?.use { inputStream ->
-            tempFile.outputStream().use { outputStream ->
-                inputStream.copyTo(outputStream)
-            }
-        }
-        return tempFile.absolutePath
-    }
-
-    fun readContentAsStringList(contentResolver: ContentResolver, contentUri: Uri): List<String> {
-        contentResolver.openInputStream(contentUri).use { inputStream ->
-            BufferedReader(InputStreamReader(inputStream)).use { reader ->
-                val lines = mutableListOf<String>()
-                var line: String? = reader.readLine()
-                while (line != null) {
-                    lines.add(line)
-                    line = reader.readLine()
-                }
-                return lines
-            }
-        }
-    }
-
     fun srtToHtml(lines: List<String>): String {
         val subtitles = mutableListOf<String>()
 
@@ -246,14 +221,6 @@ object HelperUnit {
         }
 
         return subtitles.joinToString(separator = "") { "<p>${it}</p>" }
-    }
-
-    @JvmStatic
-    fun fileName(url: String?): String {
-        val sdf = SimpleDateFormat("yyyy-MM-dd_HH-mm-ss", Locale.getDefault())
-        val currentTime = sdf.format(Date())
-        val domain = Uri.parse(url).host?.replace("www.", "")?.trim { it <= ' ' }.orEmpty()
-        return domain.replace(".", "_").trim { it <= ' ' } + "_" + currentTime.trim { it <= ' ' }
     }
 
     @JvmStatic
@@ -309,482 +276,6 @@ object HelperUnit {
         val nonNullUrl = url ?: return
         val intent = Intent(Intent.ACTION_VIEW, Uri.parse(nonNullUrl))
         activity.startActivity(Intent.createChooser(intent, title))
-    }
-
-    fun loadAssetFileToString(context: Context, filename: String): String {
-        val inputStream = context.assets.open(filename)
-        val bufferedReader = BufferedReader(InputStreamReader(inputStream))
-        val stringBuilder = StringBuilder()
-        bufferedReader.useLines { lines -> lines.forEach { stringBuilder.append(it) } }
-
-        return stringBuilder.toString()
-    }
-
-    fun getFileInfoFromContentUri(context: Context, contentUri: Uri): Pair<String?, String?> {
-        var fileName: String? = null
-        var mimeType: String? = null
-
-        val cursor = context.contentResolver.query(contentUri, null, null, null, null)
-        cursor?.use {
-            try {
-                if (it.moveToFirst()) {
-                    val displayNameColumnIndex = it.getColumnIndex(OpenableColumns.DISPLAY_NAME)
-
-                    fileName = if (displayNameColumnIndex != -1)
-                        it.getString(displayNameColumnIndex)
-                    else {
-                        contentUri.path?.split("/")?.last()
-                    }
-
-                    val mimeTypeColumnIndex = it.getColumnIndex(
-                        MimeTypeMap.getSingleton()
-                            .getExtensionFromMimeType(it.getString(it.getColumnIndexOrThrow("mime_type")))
-                    )
-                    mimeType = if (mimeTypeColumnIndex != -1)
-                        it.getString(mimeTypeColumnIndex)
-                    else {
-                        context.contentResolver.getType(contentUri)
-                    }
-                }
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
-        }
-
-        return Pair(fileName, mimeType)
-    }
-
-    suspend fun upgradeToLatestRelease(
-        context: Context,
-        progressCallback: (suspend (Float) -> Unit)? = null,
-    ) {
-        if (isAppInstalledFromPlayStore(context)) {
-            progressCallback?.invoke(1.0f)
-            withContext(Dispatchers.Main) {
-                // launch play store with my app page
-                val intent = Intent(Intent.ACTION_VIEW).apply {
-                    data =
-                        Uri.parse("https://play.google.com/store/apps/details?id=${context.packageName}")
-                    setPackage("com.android.vending")
-                }
-                context.startActivity(intent)
-            }
-            return
-        }
-
-        progressCallback?.invoke(0.1f)
-        val url = "https://api.github.com/repos/plateaukao/einkbro/releases"
-        val request = Request.Builder().url(url).build()
-
-        client.newCall(request).execute().use { response ->
-            try {
-                progressCallback?.invoke(0.2f)
-                val jsonArray = JSONArray(response.body!!.string())
-                val latestRelease = jsonArray.getJSONObject(0)
-                val tagName = latestRelease.getString("tag_name").replace("v", "")
-                val isPreRelease = latestRelease.getBoolean("prerelease")
-
-                if (tagName > BuildConfig.VERSION_NAME && !isPreRelease) {
-                    progressCallback?.invoke(0.3f)
-
-                    val downloadUrl = latestRelease.getJSONArray("assets")
-                        .getJSONObject(0)
-                        .getString("browser_download_url")
-
-                    val file = File.createTempFile("temp", ".apk", context.cacheDir)
-                    downloadFileWithProgress(downloadUrl, file.absolutePath, progressCallback, 0.4f, 0.5f)
-
-                    progressCallback?.invoke(0.9f)
-                    installApkFromFile(context, file)
-                    progressCallback?.invoke(1.0f)
-                } else {
-                    progressCallback?.invoke(1.0f)
-                    withContext(Dispatchers.Main) {
-                        EBToast.show(context, "Already up to date")
-                    }
-                }
-            } catch (e: Exception) {
-                e.printStackTrace()
-                progressCallback?.invoke(0.0f)
-                withContext(Dispatchers.Main) {
-                    EBToast.show(context, "Something went wrong")
-                }
-            }
-        }
-    }
-
-    private suspend fun installApkFromFile(context: Context, file: File) {
-        val apkUri = FileProvider.getUriForFile(
-            context,
-            BuildConfig.APPLICATION_ID + ".fileprovider",
-            file
-        )
-
-        val intent = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            @Suppress("DEPRECATION")
-            Intent(Intent.ACTION_INSTALL_PACKAGE).apply {
-                data = apkUri
-                flags = Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_ACTIVITY_NEW_TASK
-            }
-        } else {
-            Intent(Intent.ACTION_VIEW).apply {
-                setDataAndType(apkUri, "application/vnd.android.package-archive")
-                flags = Intent.FLAG_ACTIVITY_NEW_TASK
-            }
-        }
-            .apply {
-                putExtra(Intent.EXTRA_NOT_UNKNOWN_SOURCE, true)
-                putExtra(Intent.EXTRA_INSTALLER_PACKAGE_NAME, context.packageName)
-            }
-
-        withContext(Dispatchers.Main) {
-            context.startActivity(intent)
-        }
-    }
-
-    private suspend fun downloadFileWithProgress(
-        url: String,
-        destinationPath: String,
-        progressCallback: (suspend (Float) -> Unit)?,
-        baseProgress: Float,
-        progressScale: Float
-    ) {
-        val request = Request.Builder().url(url).build()
-        client.newCall(request).execute().use { response ->
-            if (!response.isSuccessful) throw IOException("Failed to download file: $response")
-
-            val responseBody = response.body ?: throw IOException("Response body is null")
-            val contentLength = responseBody.contentLength()
-            val inputStream = responseBody.byteStream()
-
-            FileOutputStream(destinationPath).use { outputStream ->
-                val buffer = ByteArray(8192)
-                var totalBytesRead = 0L
-                var bytesRead: Int
-
-                while (inputStream.read(buffer).also { bytesRead = it } != -1) {
-                    outputStream.write(buffer, 0, bytesRead)
-                    totalBytesRead += bytesRead
-
-                    if (contentLength > 0) {
-                        val progress = baseProgress + (totalBytesRead.toFloat() / contentLength.toFloat()) * progressScale
-                        progressCallback?.invoke(progress)
-                    }
-                }
-            }
-        }
-    }
-
-    suspend fun upgradeFromSnapshot(
-        context: Context,
-        progressCallback: (suspend (Float) -> Unit)? = null,
-    ) {
-        val url =
-            "https://nightly.link/plateaukao/einkbro/workflows/buid-app-workflow.yaml/main/app-arm64-v8a-release.apk.zip"
-        try {
-            progressCallback?.invoke(0.1f)
-
-            val zipFile = File.createTempFile("snapshot", ".zip", context.cacheDir)
-            downloadFileWithProgress(url, zipFile.absolutePath, progressCallback, 0.2f, 0.5f)
-
-            progressCallback?.invoke(0.7f)
-            extractApkAndInstallWithProgress(zipFile.inputStream(), context, progressCallback)
-            zipFile.delete()
-        } catch (e: Exception) {
-            e.printStackTrace()
-            progressCallback?.invoke(0.0f)
-            withContext(Dispatchers.Main) {
-                EBToast.show(context, "Something went wrong")
-            }
-        }
-    }
-
-    private suspend fun extractApkAndInstallWithProgress(
-        inputStream: InputStream?,
-        context: Context,
-        progressCallback: (suspend (Float) -> Unit)?,
-    ) {
-        val zipInputStream = ZipInputStream(inputStream)
-
-        var found = false
-        var zipEntry = zipInputStream.nextEntry
-        while (zipEntry != null) {
-            if (zipEntry.name.endsWith(".apk")) {
-                val tempFile = File("${context.cacheDir.absolutePath}/app.apk")
-                if (tempFile.exists()) {
-                    tempFile.delete()
-                }
-                tempFile.createNewFile()
-                tempFile.deleteOnExit()
-
-                val entrySize = zipEntry.size // uncompressed size, may be -1
-                FileOutputStream(tempFile).use { fos ->
-                    val buffer = ByteArray(8192)
-                    var totalBytesWritten = 0L
-                    var bytesRead: Int
-                    while (zipInputStream.read(buffer).also { bytesRead = it } != -1) {
-                        fos.write(buffer, 0, bytesRead)
-                        totalBytesWritten += bytesRead
-                        if (entrySize > 0) {
-                            val progress = 0.75f + (totalBytesWritten.toFloat() / entrySize.toFloat()) * 0.15f
-                            progressCallback?.invoke(progress)
-                        }
-                    }
-                }
-
-                progressCallback?.invoke(0.9f)
-                installApkFromFile(context, tempFile)
-                progressCallback?.invoke(1.0f)
-
-                found = true
-                break
-            }
-            zipEntry = zipInputStream.nextEntry
-        }
-        zipInputStream.closeEntry()
-        zipInputStream.close()
-
-        if (!found) {
-            throw IOException("No APK found in snapshot archive")
-        }
-    }
-
-    @Suppress("DEPRECATION")
-    private fun isAppInstalledFromPlayStore(context: Context): Boolean {
-        val packageName = context.packageName
-        val pm = context.packageManager
-
-        return try {
-            val installerPackageName = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-                pm.getInstallSourceInfo(packageName).installingPackageName
-            } else {
-                pm.getInstallerPackageName(packageName)
-            }
-            "com.android.vending" == installerPackageName
-        } catch (e: Exception) {
-            false
-        }
-    }
-
-    private val fileCache = mutableMapOf<String, String>()
-    fun loadAssetFile(fileName: String): String {
-        if (fileCache.containsKey(fileName)) {
-            return fileCache[fileName]!!
-        }
-
-        try {
-            val jsContent = EinkBroApplication.instance.assets.open(fileName).bufferedReader().use { it.readText() }
-            fileCache[fileName] = jsContent
-            return jsContent
-        } catch (e: IOException) {
-            Log.e("HelperUnit", "Failed to load asset file: $fileName")
-            e.printStackTrace()
-            return ""
-        }
-    }
-
-    private const val DEFAULT_FONT_SIZE = 18
-
-    /**
-     * Parses a given markdown text and converts it into an [AnnotatedString] with appropriate styles.
-     * from: mdparserkitcore/src/main/java/com/daksh/mdparserkit/core/ParseMarkdown.kt
-     *
-     * @param markdownText The input markdown text to parse.
-     * @return An [AnnotatedString] with styles applied according to the markdown syntax.
-     */
-    fun parseMarkdown(markdownText: String): AnnotatedString {
-        val lines = markdownText.split("\n")
-        val resultBuilder = AnnotatedString.Builder()
-        var currentStyle: SpanStyle
-
-        lines.forEach { line ->
-            when {
-                // Heading 1: Extracting content, applying bold style, and appending to resultBuilder
-                line.startsWith("# ") -> {
-                    val content = line.removePrefix("# ").trim()
-                    textMarkDown(
-                        content,
-                        resultBuilder,
-                        fontSize = (DEFAULT_FONT_SIZE + 4).sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-                // Similar processing for Heading 2 to Heading 6
-                // Heading 2
-                line.startsWith("## ") -> {
-                    val content = line.removePrefix("## ").trim()
-                    textMarkDown(
-                        content,
-                        resultBuilder,
-                        fontSize = (DEFAULT_FONT_SIZE + 3).sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-                // Heading 3
-                line.startsWith("### ") -> {
-                    val content = line.removePrefix("### ").trim()
-                    textMarkDown(
-                        content,
-                        resultBuilder,
-                        fontSize = (DEFAULT_FONT_SIZE + 2).sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-                // Heading 4
-                line.startsWith("#### ") -> {
-                    val content = line.removePrefix("#### ").trim()
-                    textMarkDown(
-                        content,
-                        resultBuilder,
-                        fontSize = (DEFAULT_FONT_SIZE + 2).sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-                // Heading 5
-                line.startsWith("##### ") -> {
-                    val content = line.removePrefix("##### ").trim()
-                    textMarkDown(
-                        content,
-                        resultBuilder,
-                        fontSize = (DEFAULT_FONT_SIZE + 1).sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-                // Heading 6
-                line.startsWith("###### ") -> {
-                    val content = line.removePrefix("###### ").trim()
-                    textMarkDown(
-                        content,
-                        resultBuilder,
-                        fontSize = (DEFAULT_FONT_SIZE + 1).sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-                // Unordered list item: Extracting content, applying bold style, appending bullet point symbol, and appending to resultBuilder
-                line.startsWith("* ") || line.startsWith("- ") -> {
-                    val content = line.removePrefix("* ").removePrefix("- ").trim()
-                    currentStyle = SpanStyle(
-                        fontWeight = FontWeight.Bold,
-                        fontSize = DEFAULT_FONT_SIZE.sp
-                    )
-                    resultBuilder.append(
-                        AnnotatedString("• ", currentStyle)
-                    )
-                    textMarkDown(content, resultBuilder, fontSize = 14.sp)
-                }
-                // Ordered list item: Extracting content, applying bold style, appending number and period, and appending to resultBuilder
-                line.matches(Regex("^\\d+\\.\\s.*$")) -> {
-                    val regex = Regex("^\\d+\\.\\s.*$")
-                    val startIndex = regex.find(line)?.range?.first ?: 0
-                    currentStyle = SpanStyle(
-                        fontWeight = FontWeight.Bold,
-                        fontSize = DEFAULT_FONT_SIZE.sp
-                    )
-                    val annotatedString = buildAnnotatedString {
-                        if (startIndex > 0) {
-                            append(line.substring(0, startIndex))
-                        }
-                        withStyle(currentStyle) {
-                            append(line.substring(startIndex, startIndex + 2))
-                        }
-                    }
-                    resultBuilder.append(annotatedString)
-                    textMarkDown(
-                        inputText = line.substring(startIndex + 2, line.length),
-                        resultBuilder = resultBuilder,
-                        fontSize = 16.sp
-                    )
-                }
-                // Remaining Text
-                else -> {
-                    textMarkDown(line, resultBuilder, fontSize = DEFAULT_FONT_SIZE.sp)
-                }
-            } // Appending new line
-            resultBuilder.append("\n")
-        }
-        return resultBuilder.toAnnotatedString().trim() as AnnotatedString
-    }
-
-    /**
-     * Converts markdown-style text formatting to [AnnotatedString] with appropriate [SpanStyle]s.
-     *
-     * @param inputText The input text to be converted.
-     * @param resultBuilder The [AnnotatedString.Builder] to append the converted text to.
-     * @param fontSize The desired font size for the text.
-     * @return The converted text with markdown formatting replaced by appropriate [SpanStyle]s.
-     */
-    private fun textMarkDown(
-        inputText: String,
-        resultBuilder: AnnotatedString.Builder,
-        fontSize: TextUnit,
-        fontWeight: FontWeight = FontWeight.Normal,
-    ) {
-        val boldItalicPattern = Regex("\\*\\*[*_](.*?)[*_]\\*\\*")
-        val boldPattern = Regex("\\*\\*(.*?)\\*\\*")
-        val italicPattern = Regex("[*_](.*?)[*_]")
-        val strikethroughPattern = Regex("~~(.+?)~~")
-
-        var currentIndex = 0
-
-        while (currentIndex < inputText.length) {
-            val nextBoldItalic = boldItalicPattern.find(inputText, startIndex = currentIndex)
-            val nextBold = boldPattern.find(inputText, startIndex = currentIndex)
-            val nextItalic = italicPattern.find(inputText, startIndex = currentIndex)
-            val nextStrikethrough = strikethroughPattern.find(inputText, startIndex = currentIndex)
-
-            val nextMarkDown = listOfNotNull(
-                nextBoldItalic,
-                nextBold,
-                nextItalic,
-                nextStrikethrough
-            ).minByOrNull { it.range.first }
-
-            if (nextMarkDown != null) {
-                if (nextMarkDown.range.first > currentIndex) {
-                    // Append any normal text before the markdown
-                    val normalText = inputText.substring(currentIndex, nextMarkDown.range.first)
-                    val style = SpanStyle(fontWeight = fontWeight, fontSize = fontSize)
-                    resultBuilder.append(AnnotatedString(normalText, style))
-                }
-
-                val matchText = nextMarkDown.groupValues.getOrNull(1).orEmpty()
-
-                val style = when (nextMarkDown) {
-                    nextBoldItalic -> SpanStyle(
-                        fontWeight = FontWeight.Bold,
-                        fontStyle = FontStyle.Italic,
-                        fontSize = fontSize
-                    )
-
-                    nextBold -> SpanStyle(
-                        fontWeight = FontWeight.Bold,
-                        fontSize = DEFAULT_FONT_SIZE.sp
-                    )
-
-                    nextItalic -> SpanStyle(
-                        fontStyle = FontStyle.Italic,
-                        fontSize = fontSize
-                    )
-
-                    nextStrikethrough -> SpanStyle(
-                        textDecoration = TextDecoration.LineThrough,
-                        fontSize = fontSize
-                    )
-
-                    else -> throw IllegalStateException("Unhandled markdown type")
-                }
-
-                resultBuilder.append(AnnotatedString(matchText, style))
-
-                currentIndex = nextMarkDown.range.last + 1
-            } else {
-                // Append any remaining text if no more markdown found
-                val normalText = inputText.substring(currentIndex)
-                val style = SpanStyle(fontWeight = fontWeight, fontSize = fontSize)
-                resultBuilder.append(AnnotatedString(normalText, style))
-                currentIndex = inputText.length
-            }
-        }
     }
 
     /**
@@ -872,4 +363,3 @@ fun String.pruneWebTitle(): String =
     } else {
         this
     }
-
